mm: Introducing dynamic buffer for debug_phys_addr
The formatted write support is required for multiple data types to be
stored in a string. Which will be used in further patches.

The buffer is dynamic in nature to store any amount of data that needs
to be stored in the file, which can be reset to page size removing all
the data by writing empty string into the file.
The buffer is originally of page size and grows if required with
twice the size of buffer each time.

The read is done in chunks of 512 bytes for complete data available.

Signed-off-by: Mukesh Kumar Chaurasiya <mchauras@linux.ibm.com>