mm: get kernel symbols for a phys addr
linux kernel gives us an interface to fetch kernel symbol from a given
virtual address. Leveraging the feature, if the memory address is
not found in any process and kmalloc space, it looks into the kernel
symbols and try to see if it stores a kernel symbol.

New macro for phys to virt for symbol is defined for x86 to get the
virtual address of the symbol.

Signed-off-by: Mukesh Kumar Chaurasiya <mchauras@linux.ibm.com>