mm: get vmap user for a phys addr
Introducing a new function to get the vmap_area for a physical address
the vmap_area contains the information of the allocator. Leveraging
this thing we iterate over all the vmap nodes and find the node matching
our physical address and extract the start address and caller from the
vmap area.

Signed-off-by: Mukesh Kumar Chaurasiya <mchauras@linux.ibm.com>